Re-organized level1 xaxpy kernel
